+++ UnitTesting/HarryPotter/HarryPotter.Tests.cs
@@ -1,69 +1,69 @@
-namespace Katas
+namespace HarryPotter
 {
-	public class HarryPotterTests
-	{
+    public class HarryPotterTests
+    {
 
-		private List<Book> books = new List<Book>()
-		{
-			new Book()
-			{
-				Name = "Harry Potter and the Philosopher's Stone",
-				Order = 1,
-				Price = 8,
-			},
-			new Book()
-			{
-				Name = "Harry Potter and the Chamber of Secrets",
-				Order = 2,
-				Price = 8,
-			},
-			new Book()
-			{
-				Name = "Harry Potter and the Prisoner of Azkaban",
-				Order = 3,
-				Price = 8,
-			},
-			new Book()
-			{
-				Name = "Harry Potter and the Goblet of Fire",
-				Order = 4,
-				Price = 8,
-			},
-			new Book()
-			{
-				Name = "Harry Potter and the Order of the Phoenix",
-				Order = 5,
-				Price = 8,
-			},
-		};
+        private List<Book> books = new List<Book>()
+        {
+            new Book()
+            {
+                Name = "Harry Potter and the Philosopher's Stone",
+                Order = 1,
+                Price = 8,
+            },
+            new Book()
+            {
+                Name = "Harry Potter and the Chamber of Secrets",
+                Order = 2,
+                Price = 8,
+            },
+            new Book()
+            {
+                Name = "Harry Potter and the Prisoner of Azkaban",
+                Order = 3,
+                Price = 8,
+            },
+            new Book()
+            {
+                Name = "Harry Potter and the Goblet of Fire",
+                Order = 4,
+                Price = 8,
+            },
+            new Book()
+            {
+                Name = "Harry Potter and the Order of the Phoenix",
+                Order = 5,
+                Price = 8,
+            },
+        };
 
-		[Fact]
-		public void CalculateCartTotal_OneBookDoesNotHaveDiscount()
-		{
-			var cartBooks = new List<Book>()
-			{
-				books[0],
-			};
-			var cart = new Cart(cartBooks);
+        [Fact]
+        public void CalculateCartTotal_OneBookDoesNotHaveDiscount()
+        {
+            var cartBooks = new List<Book>()
+            {
+                books[0],
+            };
+            var cart = new Cart(cartBooks);
 
-			double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
-			Assert.Equal(8, total, 2);
-		}
+            Assert.Equal(8, total, 2);
+        }
 
-		[Fact]
-		public void CalculateCartTotal_TwoDifferentBooksHave5PercentDiscount()
-		{
-			var cartBooks = new List<Book>()
-			{
-				books[0],
-				books[1],
-			};
-			var cart = new Cart(cartBooks);
+        [Fact]
+        public void CalculateCartTotal_TwoDifferentBooksHave5PercentDiscount()
+        {
+            var cartBooks = new List<Book>()
+            {
+                books[0],
+                books[1],
+            };
+            var cart = new Cart(cartBooks);
 
-			double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
-			Assert.Equal(15.2, total, 2);
+            Assert.Equal(15.2, total, 2);
         }
 
         [Fact]
@@ -73,11 +73,11 @@ namespace Katas
             {
                 books[0],
                 books[1],
-				books[2],
+                books[2],
             };
             var cart = new Cart(cartBooks);
 
-            double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
             Assert.Equal(21.6, total, 2);
         }
@@ -94,7 +94,7 @@ namespace Katas
             };
             var cart = new Cart(cartBooks);
 
-            double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
             Assert.Equal(25.6, total, 2);
         }
@@ -112,7 +112,7 @@ namespace Katas
             };
             var cart = new Cart(cartBooks);
 
-            double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
             Assert.Equal(30, total, 2);
         }
@@ -133,7 +133,7 @@ namespace Katas
             };
             var cart = new Cart(cartBooks);
 
-            double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
             Assert.Equal(51.6, total, 2);
         }
@@ -155,7 +155,7 @@ namespace Katas
             var cartBooks = new List<Book>();
             var cart = new Cart(cartBooks);
 
-            double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
             Assert.Equal(0, total, 2);
         }
@@ -174,7 +174,7 @@ namespace Katas
             };
             var cart = new Cart(cartBooks);
 
-            double total = cart.CalculateTotal();
+            var total = cart.CalculateTotal();
 
             Assert.Equal(36, total, 2);
         }
