+++ UnitTesting/UnitTesting.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33815.320
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Katas", "Katas\Katas.csproj", "{10EF1919-9A73-499E-B25F-D39FC7449A9C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HarryPotter", "HarryPotter\HarryPotter.csproj", "{10EF1919-9A73-499E-B25F-D39FC7449A9C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DictionaryReplacer", "DictionaryReplacer\DictionaryReplacer.csproj", "{051CF5F7-78ED-4D55-93E1-00AD930830CB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DictionaryReplacer", "DictionaryReplacer\DictionaryReplacer.csproj", "{051CF5F7-78ED-4D55-93E1-00AD930830CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
